Fix CodeGlassMini on IE.
Codeglass on IE was completely broken, or at least, dijit/form/ComboBox and dijit/form/TextBox examples wouldn't run on IE8.

Dojo's AMD's loader does not play well with document.write(), or perhaps the problem was due to the document.close() call that was occuring via a setTimeout(), possibly while dojo was in the middle of waiting for the require() resources to download.

Switched the code to set the iframe contents via frame.set("src", "javascript: ..."), like the editor does.  That's putting a lot of contents into an an attribute, but it seems to work on IE8, FF10, and chrome at least.