Raw HTML button.
Add button to show full code, so that user can copy/paste, and run locally.   It was being logged to the console, but people didn't realize.   Note that the old CodeGlass had such a button/tab.

Also removing the special styling for the Run and Collapse buttons, partly because the RUN button looks like a collapse/expand icon, partly because that styling isn't even being used on dojotoolkit.org/reference-guide, but mostly because I don't have an icon for "show source".

Plus fixing spacing.